requirements.txt
django==1.8.13  # rq.filter: >=1.8,<1.9
psycopg2==2.6.1
django-redis==4.4.2
djangorestframework==3.3.3
django-extensions==1.6.6